添加用户角色管理，修改一些细节

+++ initial_version_of_springboot/src/main/java/com/fc/test/mapper/auto/TSysRoleUserMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TSysRoleUserMapper {
-    long countByExample(TSysRoleUserExample example);
+    int countByExample(TSysRoleUserExample example);
 
     int deleteByExample(TSysRoleUserExample example);
