+++ initial_version_of_springboot/src/main/java/com/fc/test/mapper/auto/TsysPermissionRoleMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TsysPermissionRoleMapper {
-    long countByExample(TsysPermissionRoleExample example);
+    int countByExample(TsysPermissionRoleExample example);
 
     int deleteByExample(TsysPermissionRoleExample example);
